psffontop: Fix warning in assemble_utf8
warning: conversion to 'unsigned int' from 'long unsigned int' may alter its value [-Wconversion]

+++ src/psffontop.c
@@ -90,7 +90,7 @@ static unsigned int
 assemble_utf8(char **inptr, int cnt)
 {
 	int err;
-	unsigned long uc;
+	unsigned int uc;
 	char *u;
 
 	uc = from_utf8(inptr, cnt, &err);
